Update README.md
If HostingEnvironment.UnregisterObject(this) is not called at the end of the Execute() method IIS will keep a reference of the Job and so it will not be garbage collected. This can be reproduced by putting a breakpoint in the Stop(bool) method and stopping IIS (Express). The breakpoint will hit for every job that was executed.